model and template name fields combined in fields module
