+++ form_designer/fields.py
@@ -35,3 +35,35 @@ class ModelNameField(models.CharField):
         defaults = {'form_class': ModelNameFormField}
         defaults.update(kwargs)
         return super(ModelNameField, self).formfield(**defaults)
+
+class TemplateFormField(forms.CharField):
+
+    def clean(self, value):
+        """
+        Validates that the input can be compiled as a template.
+        """
+        value = super(TemplateFormField, self).clean(value)
+        from django.template import Template, TemplateSyntaxError
+        try:
+            Template(value)
+        except TemplateSyntaxError as error:
+            raise forms.ValidationError(error)
+        return value
+
+class TemplateCharField(models.CharField):
+
+    def formfield(self, **kwargs):
+        # This is a fairly standard way to set up some defaults
+        # while letting the caller override them.
+        defaults = {'form_class': TemplateFormField}
+        defaults.update(kwargs)
+        return super(TemplateCharField, self).formfield(**defaults)
+
+class TemplateTextField(models.TextField):
+
+    def formfield(self, **kwargs):
+        # This is a fairly standard way to set up some defaults
+        # while letting the caller override them.
+        defaults = {'form_class': TemplateFormField}
+        defaults.update(kwargs)
+        return super(TemplateTextField, self).formfield(**defaults)
